Conflicts:
	deps/hiredis/async.c
	deps/hiredis/examples/example-ae.c
	deps/hiredis/hiredis.c
	deps/hiredis/hiredis.h
	deps/hiredis/net.c
	deps/hiredis/net.h
	deps/jemalloc/bin/pprof
	msvs/RedisServer.vcxproj
	src/Win32_Interop/Win32_CommandLine.cpp
	src/Win32_Interop/Win32_QFork.cpp
	src/Win32_Interop/win32_wsiocp.c
	src/Win32_Interop/win32fixes.c
	src/config.c
	src/networking.c
	src/rdb.c
	src/redis-benchmark.c
	src/redis-cli.c
	src/redis.c
	src/redis.h
	tests/integration/replication-psync.tcl
	tests/unit/aofrw.tcl

+++ deps/hiredis/sds.c
@@ -324,7 +324,7 @@ int sdsll2str(char *s, long long value) {
         s++;
         p--;
     }
-    return l;
+    return (int)l;
 }
 
 /* Identical sdsll2str(), but for unsigned long long type. */
@@ -353,7 +353,7 @@ int sdsull2str(char *s, unsigned long long v) {
         s++;
         p--;
     }
-    return l;
+    return (int)l;
 }
 
 /* Create an sds string from a long long value. It is much faster than:
@@ -454,7 +454,7 @@ sds sdscatfmt(sds s, char const *fmt, ...) {
 
     va_start(ap,fmt);
     f = fmt;    /* Next format specifier byte to process. */
-    i = initlen; /* Position of the next byte to write to dest str. */
+    i = (int)initlen; /* Position of the next byte to write to dest str. */
     while(*f) {
         char next, *str;
         int l;
@@ -475,7 +475,7 @@ sds sdscatfmt(sds s, char const *fmt, ...) {
             case 's':
             case 'S':
                 str = va_arg(ap,char*);
-                l = (next == 's') ? strlen(str) : sdslen(str);
+                l = (int)((next == 's') ? strlen(str) : sdslen(str));
                 if (sh->free < l) {
                     s = sdsMakeRoomFor(s,l);
                     sh = (void*) (s-(sizeof(struct sdshdr)));
